Corrigindo sobreescrita no canvas do projeto, agora objetos colocados no painel de desenho são reaproveitados das intâncias dos objetos desenhavéis

+++ CearaGameTools/src/cgt/core/CGTProject.java
@@ -5,8 +5,8 @@ import java.io.Serializable;
 public class CGTProject implements Serializable {
 
     private String projectName;
-    private int canvasWidth;
-    private int canvasHeight;
+    private double canvasWidth;
+    private double canvasHeight;
 
     public CGTProject(String projectName) {
         this.projectName = projectName;
@@ -20,19 +20,19 @@ public class CGTProject implements Serializable {
         this.projectName = projectName;
     }
 
-    public int getCanvasWidth() {
+    public double getCanvasWidth() {
         return canvasWidth;
     }
 
-    public void setCanvasWidth(int canvasWidth) {
+    public void setCanvasWidth(double canvasWidth) {
         this.canvasWidth = canvasWidth;
     }
 
-    public int getCanvasHeight() {
+    public double getCanvasHeight() {
         return canvasHeight;
     }
 
-    public void setCanvasHeight(int canvasHeight) {
+    public void setCanvasHeight(double canvasHeight) {
         this.canvasHeight = canvasHeight;
     }
 }
